moved errorWindow to FormComponents
